Wulffrith Launcher 1.4
- Adds A Right-Click Function To All The App Icons To Edit The Apps Easier.

The launcher has gotten to a point where it is useable enough for me to be satisfied with it. Which means, this is the final update I will be doing on this project for a good while.

That said, I may get to the rest of the implementations on the roadmap but I don't see these as necessary right now and I'd like to move on to other things for the time being.